www.wireswire.com
The wire fill factor is defined as the ratio of the amount of conductor material copper (without insulation) to the available winding space lot, and it has a significant impact on the power density, thermal conductivity, manufacturability, and overall cost of a motor. Fill rate is directly related to the current density limit, which is related to torque and power. The cross-section of enameled wire defines two major categories as round wire and rectangular wire. The inherent characteristic of the round shape is that the filling factor is low, and the rectangle matches the groove shape well. It is characterized by a high filling factor. . Due to their rectangular cross-section, the fill factor can reach values ​​of 80-85% in machines using them.

Compared with the traditional round enameled wire motor, the winding motor with rectangular cross-section can increase the slot full rate by 20%~30%, and the copper ratio per unit cross-sectional area is higher, which effectively reduces the winding resistance and thus reduces the copper loss. The thicker and smaller the resistance, the less energy will be lost on the wire due to heat generation. The length of the motor end of the rectangular cross-section is 5~10mm shorter than that of the round wire, which effectively reduces the copper consumption of the end winding and further improves the motor efficiency. The rectangular cross-section motor has a higher slot full rate. Under the same volume, the rectangular cross-section winding motor has a larger effective stator winding area, and the rectangular cross-section winding motor can output higher power and torque under the same loss.

The rectangular cross-section winding motor flat wire has a more regular shape, fits closely in the stator slot, and has better contact with the stator core teeth and yoke, reducing the thermal resistance in the slot and improving the heat conduction efficiency. The motor using the flat enameled wire structure has better rigidity, and the whole machine will also have better rigidity. The flat enameled wire winding is inserted through the end of the iron core, and does not need to be inserted from the slot. In the electromagnetic design, a smaller slot design can be selected to effectively reduce the cogging torque pulsation.